[Time] 20-Feb-2020 | 14:25
[Description]

[Watch][ChangeLog]
1. Refactoring Modules

[Phone][ChangeLog]
2. Popup AppShortcut Added
1. Refactoring Modules

[Build][ChangeLog]
Adding Initial Dependencies

+++ GeekyGifyPhone/src/main/java/net/geeksempire/geeky/gify/Utils/Networking/DownloadGif.kt
@@ -1,8 +1,8 @@
 /*
  * Copyright © 2020 By Geeks Empire.
  *
- * Created by Elias Fazel on 2/19/20 5:35 PM
- * Last modified 2/19/20 5:30 PM
+ * Created by Elias Fazel on 2/20/20 2:24 PM
+ * Last modified 2/20/20 12:47 PM
  *
  * Licensed Under MIT License.
  * https://opensource.org/licenses/MIT
@@ -11,6 +11,7 @@
 package net.geeksempire.geeky.gify.Utils.Networking
 
 import android.content.Context
+import android.util.Log
 import kotlinx.coroutines.*
 import net.geeksempire.geeky.gify.Utils.LinkInformation.ExtractLinkData
 import java.io.File
@@ -21,24 +22,31 @@ class DownloadGif (var context: Context) {
 
     @Throws(Exception::class)
     fun downloadGifFile(linkToGif: String) : Deferred<File> = CoroutineScope(SupervisorJob() + Dispatchers.Default).async {
+        Log.d(this@DownloadGif.javaClass.simpleName, linkToGif)
 
-        val downloadGifByte = URL(linkToGif).readBytes()
-
-        File(context.externalMediaDirs[0].path + File.separator + "SharedGifCollection").let {
+        val geekyGifyCollectionFolderPath = context.externalMediaDirs[0].path + File.separator + "SharedGifCollection"
+        File(geekyGifyCollectionFolderPath).let {
             if (!it.exists()) {
                 it.mkdirs()
             }
         }
 
         val gifId = ExtractLinkData().extractGifId(linkToGif)
-        val filePath = context.externalMediaDirs[0].path + File.separator + "SharedGifCollection" + File.separator + "GeekyGify${gifId}" + ".GIF"
+        val filePath = geekyGifyCollectionFolderPath + File.separator + "GeekyGify${gifId}" + ".GIF"
         val gifFile = File(filePath)
-        val fileOutputStream = FileOutputStream(gifFile)
 
-        fileOutputStream.write(downloadGifByte)
+        if (gifFile.exists()) {
 
-        fileOutputStream.flush()
-        fileOutputStream.close()
+        } else {
+            val downloadGifByte = URL(linkToGif).readBytes()
+
+            val fileOutputStream = FileOutputStream(gifFile)
+
+            fileOutputStream.write(downloadGifByte)
+
+            fileOutputStream.flush()
+            fileOutputStream.close()
+        }
 
         gifFile
     }
